// frontend/src/app/types/sourceFiles.ts
export interface TreeFileDto {
    id: number;
    name: string;
    originalName: string;
    fileSizeLabel: string;
    fileSizeBytes: number;
    format: string;
    statut: string;
    contactCount: number;
    listNumber: number;
    uploadedAt: string;
    isActive: boolean;
}